% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{BootstrapData}
\alias{BootstrapData}
\title{Generate Bootstrap Data}
\usage{
BootstrapData(
  Y,
  Yhat,
  bootstrap_resampling_method = c("homoscedatic", "wild.continuous", "wild.twopoint"),
  bootstrap_samples = 100
)
}
\arguments{
\item{Y}{Numeric vector of learning responses}

\item{Yhat}{Numeric vector of fitted values for learning data}

\item{Resampling.Method}{Character string specifying resampling method:
"homoscedatic", "wild.continuous", or "wild.twopoint"}

\item{NB}{Integer number of bootstrap samples (default: 100)}
}
\value{
Matrix of bootstrap samples (n x NB)
}
\description{
Generates bootstrap samples using various resampling methods
}
\examples{
\dontrun{
  Y <- rnorm(100)
  Yhat <- rnorm(100)
  W <- BootstrapData(Y, Yhat, "homoscedatic", 100)
}
}
