% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_learn.R
\name{FuNopaCl}
\alias{FuNopaCl}
\title{Functional Nonparametric Classification Learning}
\usage{
FuNopaCl(X, classes, semimetric = "Deriv", semimetric.params)
}
\arguments{
\item{X}{Matrix of functional data (each row is an observation)}

\item{classes}{Integer vector of class labels}

\item{semimetric}{Character string specifying semimetric type (default: "Deriv")}

\item{semimetric.params}{List of parameters for the semimetric}
}
\value{
Object of class FuNopaCl containing:
  \itemize{
    \item Semimetric: Name of the semimetric used
    \item semimetric.params: Parameters used for the semimetric
    \item k.opt: Optimal k values from cross-validation
    \item mse.learn: Learning error rate
    \item classes.estimated: Estimated classes for learning set
    \item Prob.estimated: Probability matrix for learning set
    \item X.learn: Learning data
    \item classes.learn: Learning class labels
  }
}
\description{
Performs nonparametric classification for functional data using k-NN
with local cross-validation
}
\examples{
\dontrun{
  # Example usage
  X <- matrix(rnorm(100 * 50), 100, 50)
  classes <- rep(1:2, each = 50)
  params <- list(q = 2, nknot = 10, range.grid = c(0, 1))
  model <- FuNopaCl(X, classes, "Deriv", params)
}
}
