% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr_learn.R
\name{FuNopaRe}
\alias{FuNopaRe}
\title{Functional Nonparametric Regression Learning}
\usage{
FuNopaRe(X, Y, semimetric = "Deriv", semimetric.params, bandwidth = "CV")
}
\arguments{
\item{X}{Matrix of functional data (each row is an observation)}

\item{Y}{Numeric vector of response values}

\item{semimetric}{Character string specifying semimetric type (default: "Deriv")}

\item{semimetric.params}{List of parameters for the semimetric}

\item{bandwidth}{Character string specifying bandwidth selection method:
"CV" for cross-validation, "kNNgCV" for k-NN global CV, "kNNlCV" for k-NN local CV}
}
\value{
Object of class FuNopaRe containing:
  \itemize{
    \item Semimetric: Name of the semimetric used
    \item Method: Method used for prediction
    \item semimetric.params: Parameters used for the semimetric
    \item h.opt or k.opt: Optimal bandwidth or k value
    \item mse.learn: Learning mean squared error
    \item Y.hat: Fitted values
    \item X.learn: Learning data
    \item Y.learn: Learning response values
  }
}
\description{
Performs nonparametric regression for functional data using kernel methods
}
\examples{
\dontrun{
  # Example usage
  X <- matrix(rnorm(100 * 50), 100, 50)
  Y <- rnorm(100)
  params <- list(q = 2, nknot = 10, range.grid = c(0, 1))
  model <- FuNopaRe(X, Y, "Deriv", params, "kNNgCV")
}
}
