% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelClassificationkNN}
\alias{KernelClassificationkNN}
\title{Kernel Classification with k-NN}
\usage{
KernelClassificationkNN(DistanceMatrix, classes, neighbours)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances (n x m)}

\item{classes}{Integer vector of class labels}

\item{neighbours}{Integer vector of neighbour counts}
}
\value{
List containing classes.predicted and Prob.predicted
}
\description{
Performs kernel classification using k-nearest neighbors
}
