% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelClassificationkNNlCV}
\alias{KernelClassificationkNNlCV}
\title{Kernel Classification with k-NN and Local Cross-Validation}
\usage{
KernelClassificationkNNlCV(DistanceMatrix, classes, knnlen)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances (square, n x n)}

\item{classes}{Integer vector of class labels}

\item{knnlen}{Integer maximum number of neighbours to consider}
}
\value{
List containing kopt, classes.estimated, and Prob.estimated
}
\description{
Performs kernel classification with local cross-validation for k selection
}
