% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPrediction}
\alias{KernelPrediction}
\title{Kernel Prediction with Fixed Bandwidth}
\usage{
KernelPrediction(DistanceMatrix, response, bandwidth)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances}

\item{response}{Numeric vector of responses}

\item{bandwidth}{Numeric bandwidth parameter}
}
\value{
Numeric vector of predictions
}
\description{
Performs kernel prediction using a fixed bandwidth
}
