% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPredictionBoot}
\alias{KernelPredictionBoot}
\title{Kernel Prediction with Bootstrap}
\usage{
KernelPredictionBoot(
  DistanceMatrixPred,
  YLearn,
  predresponse,
  BootMat,
  neighbours
)
}
\arguments{
\item{DistanceMatrixPred}{Numeric matrix of distances (n x m)}

\item{YLearn}{Numeric vector of learning responses}

\item{predresponse}{Numeric vector of predicted responses}

\item{BootMat}{Numeric matrix of bootstrap samples}

\item{neighbours}{Integer number of neighbours}
}
\value{
List containing pred, mu, sigma, and Mse
}
\description{
Performs kernel prediction with bootstrap confidence intervals
}
