% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPredictionCV}
\alias{KernelPredictionCV}
\title{Kernel Prediction with Cross-Validation}
\usage{
KernelPredictionCV(DistanceMatrix, response, bandwidth)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances (square, n x n)}

\item{response}{Numeric vector of responses}

\item{bandwidth}{Initial bandwidth value}
}
\value{
List containing hseq, mse, hopt, and yhat
}
\description{
Performs kernel prediction with cross-validation for bandwidth selection
}
