% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPredictionkNN}
\alias{KernelPredictionkNN}
\title{Kernel Prediction with k-Nearest Neighbors}
\usage{
KernelPredictionkNN(DistanceMatrix, response, neighbours, local)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances}

\item{response}{Numeric vector of responses}

\item{neighbours}{Integer or vector of neighbour counts}

\item{local}{Logical indicating local or global bandwidth selection}
}
\value{
Numeric vector of predictions
}
\description{
Performs kernel prediction using k-nearest neighbors
}
