% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPredictionkNNgCV}
\alias{KernelPredictionkNNgCV}
\title{Kernel Prediction with k-NN and Global Cross-Validation}
\usage{
KernelPredictionkNNgCV(DistanceMatrix, response, knnlen)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances (square, n x n)}

\item{response}{Numeric vector of responses}

\item{knnlen}{Integer maximum number of neighbours to consider}
}
\value{
List containing kopt, mse, and yhat
}
\description{
Performs kernel prediction with global cross-validation for k selection
}
