% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelPredictionkNNlCV}
\alias{KernelPredictionkNNlCV}
\title{Kernel Prediction with k-NN and Local Cross-Validation}
\usage{
KernelPredictionkNNlCV(DistanceMatrix, response, knnlen)
}
\arguments{
\item{DistanceMatrix}{Numeric matrix of distances (square, n x n)}

\item{response}{Numeric vector of responses}

\item{knnlen}{Integer maximum number of neighbours to consider}
}
\value{
List containing kopt, yhat, and mse
}
\description{
Performs kernel prediction with local cross-validation for k selection
}
