% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfda-package.R
\docType{package}
\name{nfda-package}
\alias{nfda}
\alias{nfda-package}
\title{nfda: Nonparametric Functional Data Analysis}
\description{
Nonparametric methods for functional data analysis including kernel regression and classification. Implements methods based on semimetrics (PCA, derivatives, PLS) with various bandwidth selection techniques. Originally based on methods by Ferraty and Vieu.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sipemu/nonparametric-functional-data-analysis}
  \item Report bugs at \url{https://github.com/sipemu/nonparametric-functional-data-analysis/issues}
}

}
\author{
\strong{Maintainer}: Simon Mueller \email{smuller@data-zoo.de}

}
\keyword{internal}
