% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_predict.R
\name{predict.FuNopaCl}
\alias{predict.FuNopaCl}
\title{Predict Method for Functional Nonparametric Classification}
\usage{
\method{predict}{FuNopaCl}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class FuNopaCl from FuNopaCl function}

\item{newdata}{Matrix of new functional data to predict}

\item{...}{Additional arguments (not currently used)}
}
\value{
Updated FuNopaCl object with additional fields:
  \itemize{
    \item classes.pred: Predicted class labels
    \item Prob.pred: Probability matrix for predictions
  }
}
\description{
Predicts class labels for new functional data using a trained FuNopaCl model
}
\examples{
\dontrun{
  # Example usage
  X <- matrix(rnorm(100 * 50), 100, 50)
  classes <- rep(1:2, each = 50)
  params <- list(q = 2, nknot = 10, range.grid = c(0, 1))
  model <- FuNopaCl(X, classes, "Deriv", params)
  
  X_new <- matrix(rnorm(10 * 50), 10, 50)
  predictions <- predict(model, X_new)
}
}
