% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semimetric.R
\name{Semimetric}
\alias{Semimetric}
\title{Semimetric Wrapper Function}
\usage{
Semimetric(Data1, Data2, semimetric, semimetric.params)
}
\arguments{
\item{Data1}{Matrix of functional data (first dataset)}

\item{Data2}{Matrix of functional data (second dataset)}

\item{semimetric}{Character string specifying the semimetric type}

\item{semimetric.params}{List of parameters for the semimetric}
}
\value{
List containing semimetric distance matrix and additional outputs
}
\description{
Main wrapper function for computing semimetric distances between functional data
}
