% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semimetric.R
\name{semimetric_deriv_wrapper}
\alias{semimetric_deriv_wrapper}
\title{Semimetric Based on Derivatives}
\usage{
semimetric_deriv_wrapper(
  Data1,
  Data2,
  q = 2,
  nknot = 20,
  range.grid = c(0, 1),
  Hhalf = NULL
)
}
\arguments{
\item{Data1}{Matrix of functional data (first dataset)}

\item{Data2}{Matrix of functional data (second dataset)}

\item{q}{Integer order of derivative (default: 2)}

\item{nknot}{Integer number of knots for B-spline (default: 20)}

\item{range.grid}{Numeric vector of length 2 specifying grid range (default: c(0,1))}

\item{Hhalf}{Matrix H-half from previous computation (default: NULL)}
}
\value{
List containing semimetric distance matrix and Hhalf matrix
}
\description{
Computes semimetric distance based on L2-norm of qth derivatives
}
