% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semimetric.R
\name{semimetric_pca_wrapper}
\alias{semimetric_pca_wrapper}
\title{Semimetric Based on PCA}
\usage{
semimetric_pca_wrapper(Data1, Data2, q, EigenVec = NULL)
}
\arguments{
\item{Data1}{Matrix of functional data (first dataset)}

\item{Data2}{Matrix of functional data (second dataset)}

\item{q}{Integer number of principal components}

\item{EigenVec}{Matrix of eigenvectors from previous computation (default: NULL)}
}
\value{
List containing semimetric distance matrix and eigenvector matrix
}
\description{
Computes semimetric distance based on principal component analysis
}
