% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SemimetricDeriv}
\alias{SemimetricDeriv}
\title{Semimetric Based on Derivatives}
\usage{
SemimetricDeriv(Hhalf, Bspline, Data1, Data2, twodatasets)
}
\arguments{
\item{Hhalf}{Numeric matrix H-half from SemimetricDerivDesign}

\item{Bspline}{Numeric matrix of B-spline basis}

\item{Data1}{Numeric matrix of functional data (first dataset)}

\item{Data2}{Numeric matrix of functional data (second dataset)}

\item{twodatasets}{Logical indicating if two datasets are used}
}
\value{
Numeric matrix of semimetric distances
}
\description{
Computes semimetric distance based on functional derivatives
}
\keyword{internal}
