% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SemimetricDerivDesign}
\alias{SemimetricDerivDesign}
\title{Semimetric Derivative Design Matrix}
\usage{
SemimetricDerivDesign(BsplineDeriv, weightgauss, span)
}
\arguments{
\item{BsplineDeriv}{Numeric matrix of B-spline derivatives}

\item{weightgauss}{Numeric vector of Gaussian quadrature weights}

\item{span}{Numeric span of the grid}
}
\value{
Numeric matrix H-half matrix
}
\description{
Calculates design matrix for derivative-based semimetric
}
\keyword{internal}
