% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SemimetricPCA}
\alias{SemimetricPCA}
\title{Semimetric Based on PCA}
\usage{
SemimetricPCA(Eigenvec, Data1, Data2, twodatasets)
}
\arguments{
\item{Eigenvec}{Numeric matrix of eigenvectors}

\item{Data1}{Numeric matrix of functional data (first dataset)}

\item{Data2}{Numeric matrix of functional data (second dataset)}

\item{twodatasets}{Logical indicating if two datasets are used}
}
\value{
Numeric matrix of semimetric distances
}
\description{
Computes semimetric distance based on principal component analysis
}
\keyword{internal}
