% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SemimetricPCAEV}
\alias{SemimetricPCAEV}
\title{Semimetric PCA Eigenvector Calculation}
\usage{
SemimetricPCAEV(Data1, q)
}
\arguments{
\item{Data1}{Numeric matrix of functional data}

\item{q}{Integer number of principal components}
}
\value{
Numeric matrix of eigenvectors
}
\description{
Calculates eigenvectors for PCA-based semimetric
}
\keyword{internal}
