% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semimetric.R
\name{SemimetricPLS}
\alias{SemimetricPLS}
\title{Semimetric Based on PLS}
\usage{
SemimetricPLS(Response, Data1, Data2, q)
}
\arguments{
\item{Response}{Numeric vector of responses}

\item{Data1}{Matrix of functional data (first dataset)}

\item{Data2}{Matrix of functional data (second dataset)}

\item{q}{Integer number of PLS components}
}
\value{
Matrix of semimetric distances
}
\description{
Computes semimetric distance based on partial least squares
}
