% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{BootstrapData}
\alias{BootstrapData}
\title{Generate Bootstrap Data}
\usage{
BootstrapData(
  Y.learn,
  Y.learn.hat,
  Resampling.Method = "homoscedatic",
  NB = 100
)
}
\arguments{
\item{Y.learn}{Numeric vector of learning responses}

\item{Y.learn.hat}{Numeric vector of fitted values for learning data}

\item{Resampling.Method}{Character string specifying resampling method:
"homoscedatic", "wild.continuous", or "wild.twopoint"}

\item{NB}{Integer number of bootstrap samples (default: 100)}
}
\value{
Matrix of bootstrap samples (n x NB)
}
\description{
Generates bootstrap samples using various resampling methods
}
\examples{
\dontrun{
  Y <- rnorm(100)
  Y.hat <- rnorm(100)
  W <- BootstrapData(Y, Y.hat, "homoscedatic", 100)
}
}
