% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_learn.R
\name{FuNopaCl}
\alias{FuNopaCl}
\title{Functional Nonparametric Classification Learning}
\usage{
FuNopaCl(X, Y, semimetric, k_nearest_neighbors = 20L)
}
\arguments{
\item{X}{Matrix of functional data (each row is an observation)}

\item{Y}{Integer vector of class labels}

\item{semimetric}{Semimetric object}

\item{k_nearest_neighbors}{Integer specifying the number of nearest neighbors}
}
\value{
Object of class FuNopaCl containing:
  \itemize{
    \item semimetric: Semimetric object
    \item k_opt: Optimal k values from cross-validation
    \item mse: Learning error rate
    \item Yhat: Estimated response values for learning set
    \item prob_estimated: Probability matrix for learning set
    \item X: Learning data
    \item Y: Learning response values
  }
}
\description{
Performs nonparametric classification for functional data using k-NN
with local cross-validation
}
\examples{
\dontrun{
  # Example usage
  X <- matrix(rnorm(100 * 50), 100, 50)
  Y <- rep(1:2, each = 50)
  semimetric <- semiemetric_deriv$new(q = 2, nknot = 10, range.grid = c(0, 1))
  model <- FuNopaCl(X, Y, semimetric, k_nearest_neighbors)
}
}
